refactoring of payment system

+++ src/main/java/com/example/my_book_shop_app/struct/user/UserEntity.java
@@ -28,7 +28,7 @@ public class UserEntity {
     private LocalDateTime regTime;
 
     @Column(columnDefinition = "INT NOT NULL")
-    private int balance;
+    private Double balance;
 
     @Column(columnDefinition = "VARCHAR(255) NOT NULL")
     private String name;
